run_localGPT.py updated
- Moved all model loading code out of the run_localGPT.py file.
- Removed llamacpp from requirement.txt file. It needs to be installed separately to ensure it supports GPU
